#9 Experimental UI update
First round of visual updates to the Meter Monitor application. Changes
layout from full-screen/liquid layout to a fixed-width layout, tables
stretch the full width of the layout, more breathing room has been
added to table data, table rows highlight when a mouse hovers over
them, a new nabber has been added, as well as an SVG image of the
City’s logo, and a blurred photo of downtown for a background image.